Restructured folders in the server project
