+++ HikingQuests.Server/Controllers/QuestManagementController.cs
@@ -1,5 +1,6 @@
-﻿using HikingQuests.Server.Constants;
-using HikingQuests.Server.Dtos;
+﻿using HikingQuests.Server.Application.Dtos;
+using HikingQuests.Server.Constants;
+using HikingQuests.Server.Domain.Entities;
 using HikingQuests.Server.Models;
 using Microsoft.AspNetCore.Mvc;
