+++ HikingQuests.Server/Program.cs
@@ -1,3 +1,5 @@
+using HikingQuests.Server.Domain;
+using HikingQuests.Server.Domain.Entities;
 using HikingQuests.Server.Models;
 using HikingQuests.Server.Utilities;
